Move handling of game end into state
