+++ server/src/game/model.ts
@@ -149,6 +149,10 @@ export class GameManager {
 
     const gameEndResult = checkForGameEnd(snapshot);
     if (gameEndResult.isEnded) {
+      this.update(game => {
+        game.endReason = gameEndResult.reason;
+      })
+
       this.io.emit(
         ServerEvent.GAME_OVER,
         this.gameId,
